requirements.txt
numpy
Pillow
matplotlib
ipython
ipywidgets
scikit-image
scipy
plotly